+++ jtsx.el
@@ -6,7 +6,7 @@
 ;; Maintainer: Loïc Lemaître <loic.lemaitre@gmail.com>
 ;; URL: https://github.com/llemaitre19/jtsx
 ;; Package-Requires: ((emacs "29.1"))
-;; Version: 0.5.0pre1
+;; Version: 0.5.0
 ;; Keywords: languages
 
 ;; This file is NOT part of GNU Emacs.